requirements.txt
streamlit==1.20.0
pandas==1.5.3
numpy==1.24.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
seaborn==0.12.2
pickleshare==0.7.5
xgboost==1.7.5
Pillow==9.4.0
scikit-learn==1.2.2
